seperated bot from lfg logic.
