+++ package.json
@@ -12,7 +12,9 @@
     },
     "dependencies": {
         "discord.js": "^11.3.2",
-        "dotenv": "^6.0.0"
+        "dotenv": "^6.0.0",
+        "express": "^4.16.3",
+        "redis": "^2.8.0"
     },
     "license": "MIT"
 }
